refactor: Move github-related code into packages
- Move config to pkg/config
- Create auth package for JWT and client handling
- Create agent package for agent management
- Split webhook server into server and handlers
- Improve error handling and logging
- Add context support for graceful shutdown
- Improve code organization and maintainability
